requirements/base.txt
# Core framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.9
alembic==1.13.1

# Analysis
rapidfuzz==3.5.2
scikit-learn>=1.3.0
numpy==1.26.3

# LLM
google-generativeai==0.3.2

# External APIs
httpx>=0.28.1,<1.0.0
requests==2.31.0
pyairtable==2.3.0

# Charts
plotly==5.18.0
kaleido==0.2.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
click==8.1.7
tomli==2.0.1
